streamlined element_type constructor in preparation for rewrite of get_dtype/resolve_dtype

+++ pdtypes/types/base.pxd
@@ -7,5 +7,5 @@ cdef class ElementType:
         tuple subtypes
         object atomic_type
         object extension_type
-        long hash
+        long long hash
         str slug
