+++ pdtypes/types/datetime.pyx
@@ -106,7 +106,7 @@ cdef class NumpyDatetime64Type(DatetimeType):
         bint is_categorical = False,
         bint is_sparse = False,
         str unit = None,
-        unsigned long step_size = 1
+        unsigned long long step_size = 1
     ):
         self.is_categorical = is_categorical
         self.is_sparse = is_sparse
